+++ cuda/src/utilities/utilities.cu
@@ -1,7 +1,7 @@
 
 #include "utilities.cuh"
 
-__host__ __device__ unsigned int idx(node i, node j, unsigned int n)
+__host__ __device__ size_t idx(node i, node j, size_t n)
 {
     return i*n + j;
 }
